MB-25908 json state changes to self in a tight loop
This is the first of the encoding/json improvements (and I realize it is
a bit on the large side): we are pushing down track of the scan from the
decoder and decodeState to the scanner, where it belongs.
This way when we need to change state to self, we can do a tight loop,
essentially switching to a no state change.

For small jsons this doesn't buy us a lot, but for a 10k json, my tests
indicate that we improve execution time by 5% in the initial Validate()
and another 5% in the Find() and Unmarshall() calls.

The bulk of the changes are to push down the data pointer and scan
position.

scanner.go has a new skipSpaces() scanner method to traverse blanks in
one go and stateInString() will loop through all uniniteresting 
characters in one go.

Change-Id: I0e5c986d70726a6597bb97feb27aff95eef82170
Reviewed-on: http://review.couchbase.org/83819
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>