MB-25907 improve performance of parsed values
This is the first cut of the effort to make parsed values more
performant.

We have two new Find functions: FirstFind() and FirstFindWithState()
They find first level fields in a single pass (as opposed to the two or
three passes dustin/go_jsonpointer and encoding/json do), and the
second saves all the fields scanned allong the way, making it
truly single pass.

We have two methods, because values can be accessed concurrently by 
multiple execution threads (this is particularly true of unnest which
will pull fields out and assemble multiple documents out of them, 
as the projection is assembling results) - how these are used will
become clear on the query side of things.

I have tried pooling just about anything - it's a hindrance.
Also - there is still more work to be done to make these functions
completely single pass.
Change-Id: Icf5566efb20d9f612cd90ed9c5a7c1cc04a3f87a
Reviewed-on: http://review.couchbase.org/84315
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>