MB-32158 remove pools from go_json
The current slowdown on Q1 on triton is due to the test rig sending an array of credentials and
encoding/json:Unmarshal() piling up on the array pool mutex.
Since we are not pooling arrays and objects any more, might just as well remove the pools altogether.
We will try to pool arrays and objects as part of value/parsed.go anyway.

Change-Id: I2d5d1dd920eb39df9c0c0b8cd2a1d199d4579633
Reviewed-on: http://review.couchbase.org/102681
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>