MB-67849 Proper check for int64 overflow
This bug is a regression from the fix for MB-51629/MB-54941.
Prior to these fixes we used to check number of digits not
exceeding 19 for int64. The original issue with MB-51629 is
with a negative number, where the negative sign also counts
towards the number of digits and thus we fall back to using
float64 prematurely. The new check for int64 overflow was
changed to multiplying the number and check for negative
number (which is a sign of overflow). This was subsequently
changed again with the fix for MB-54941. However, the new
check introduced for MB-54941 is also not correct for all
numbers, e.g. for the number used in this bug, it did not
properly detect the overflow.

To fix this, we revert the code back to checking number
of digits as an indication of overflow, but add new logic
to account for negative number as well, in effect, refix
the original issue of MB-51629. Also added logic to make
sure when at 19 digits, we do not go over MaxInt64.

Change-Id: Ie349bdca729daec203d67317bab5b0dd5bbef402
Reviewed-on: https://review.couchbase.org/c/go_json/+/231553
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>